src/gerenciador_energia/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
django-pandas==0.6.7
djangorestframework==3.15.2
executing==2.1.0
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
plotly==5.24.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
tenacity==9.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
